python_download_speed_check/requirements.txt
aiobotocore==0.10.2
aiofiles==0.4.0
aiohttp==3.5.4
aiomultiprocess==0.5.0
async-generator==1.10
async-timeout==3.0.1
asyncio-executor==0.0.4
asyncio-pool==0.4.0
attrs==19.1.0
boto==2.49.0
boto3==1.9.164
botocore==1.12.164
chardet==3.0.4
docutils==0.14
idna==2.8
jmespath==0.9.4
multidict==4.5.2
python-dateutil==2.8.0
python-dotenv==0.10.3
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
wrapt==1.11.1
yarl==1.3.0
